bug fixing

+++ web/protected/pages/Order/Print/OrderPrintController.php
@@ -103,7 +103,7 @@ class OrderPrintController extends BPCPageAbstract
 		$html = $this->_getPaymentSummaryRow('Total Excl. GST:', '$' . number_format($totalNoGST, 2, '.', ','), 'grandTotalNoGST');
 		$html .= $this->_getPaymentSummaryRow('Total GST:', '$' . number_format($gst, 2, '.', ','), 'gst');
 		$html .= $this->_getPaymentSummaryRow('Sub Total Incl. GST:', '$' . number_format($total, 2, '.', ','), 'grandTotal');
-		$html .= $this->_getPaymentSummaryRow('Shipping Incl. GST:', '$' . number_format(StringUtilsAbstract::getValueFromCurrency(implode('', $this->order->getInfo(OrderInfoType::ID_SHIPPING_EST_COST))), 2, '.', ','), 'grandTotal');
+		$html .= $this->_getPaymentSummaryRow('Shipping Incl. GST:', '$' . number_format((double)StringUtilsAbstract::getValueFromCurrency(implode('', $this->order->getInfo(OrderInfoType::ID_SHIPPING_EST_COST))), 2, '.', ','), 'grandTotal');
 		$html .= $this->_getPaymentSummaryRow('Paid to Date:', '$' . number_format($this->order->getTotalPaid(), 2, '.', ','), 'paidTotal');
 		$overDueClass = $this->order->getTotalDue() > 0 ? 'overdue' : '';
 		$html .= $this->_getPaymentSummaryRow('Balance Due:', '$' . number_format($this->order->getTotalDue(), 2, '.', ','), 'dueTotal ' . $overDueClass);
